분자, 분모의 접근타입을 변경하라
* long -> int 생성자와 동일한 규격으로 변경

+++ src/main/java/algorythm/Pro_1_2_16.java
@@ -113,12 +113,12 @@ class Pro_1_2_16 {
       return numerator + "/" + denominator;
     }
 
-    long getNumerator() {
-      return numerator;
+    int getNumerator() {
+      return (int) numerator;
     }
 
-    long getDenominator() {
-      return denominator;
+    int getDenominator() {
+      return (int) denominator;
     }
 
     /**
